Add shields, engine particles and radar-map
The shields are visual effect only, they are used to indicate when the ship takes damage. The shields gain a red color when the player drops to low health.

The engine particles are visual effects only. They blast off whenever the player boosts the ship.

The radar adds an extra camera to the map that renders in the GUI as an overlay. It uses a seperate render-layer to make objects appear radar-like.
A GUI-mask has been added to shape the radar as it appears on the GUI.
Note 1: every game object was given an alternate mesh and texture for render in the Radar-layer. The player's and bullet's radar-mesh was enlarged to be visible.

Note 2: It is wise to toggle off the Radar layer from the editor-view when not needed.