Add a bomb weapon - increase firerate - make shield provide protection
Added a bomb that will deal massive damage to asteroids, but will also push them around, increasing the risk factor.
Increased the firerate to make the game a bit quicker.
Made it so the player can't take aditional damage while the shield are active.
Inclreased the duration of the shield to 3 secunds.